Some changes in tokenomics
INITIAL TOTAL SUPPLY – 99 million AMG

Note:

Team-AMG: Share for the project team. The tokens are issued progressively and periodically to members over a period of 3 years.

Strategic Partners: Share for the AMG’s partner firms and cryptographic initiatives..

AMG is a deflationary token, which implies that the supply in circulation continuously dropping over time through token burning. This is paired with the inability to produce fresh tokens. More information about the total supply in circulation in real-time as well as the different carriers is available on the BSCScan